Set tab size in text editor to match default HTML
The default HTML in a new site’s `index.html` and `not_found.html` has 2-space indents. Setting the editor to match that makes the page easier to edit, because the user can use the Tab key for indentation.

This change is untested. The API call was copied from https://github.com/ajaxorg/ace/wiki/Embedding-API.