redesign
included all dev and production files for new redesign.

the index file has a different layout than the interior page files.

for more in-depth look into how these are set up, look in the templates > includes & layouts folders.

I used .liquid files to create these pages, and all the content for the files are in there, with the base layout stuff in the includes and layout files.

check the converted-html folder to see how everything should look.

(test it on an android phone.  should look good now) :)