Make "clouds" default text editor
The default solarized theme is hard to read and very dark, making it somewhat imposing for beginners, which I've seen with my students. It also doesn't go very well with the rest of the Neocities UI.

As a designer, I would recommend changing the default theme to Clouds, which is what this pull request does, but frankly almost anything is better. I also added a subtle border around the editor in order to differentiate it from the rest of the page when a light theme is selected.