Fix tempfile returning nil for empty files
This is an "monkey patch" to have Tempfile return 0 instead of nil when
a tempfile is empty. This has caused an unexplained exception in our
logs for years, and I was really surprised when I caught it, and more
surprised that nobody has ever reported it to us before.

I really think this is a bug, and have filed it as a bug on the ruby
tracker. I haven't seen any consequences (yet) to doing this, so I'm
just going to run with it and see what happens. If it blows
something up (how? what scenario?), I'll try a different approach.