Fix Mastodon Share
Fix two issues with the Mastodon share link. Mastodon was not given a
new line unlike all the other share links, so it was showing up right
next to Tumblr. Second, the code that generated query params seems to
have been copied from Tumblr or Facebook without properly testing it and
does not fill in the textbox correctly. Fix this so the prefilled "Toot"
matches what Twitter share would post.