Make ACME challenge cleanup more specific
The .well-known directory has uses other than ACME challenge response,
so deleting the whole thing as cleanup is incorrect.

I ran into an issue where files I uploaded inside that directory (which
worked initially) disappeared after a while, I assume around the time
neocities renewed the certs for my custom domain. The file (and the
.well-known dir) still appeared in GETs on the /api/list endpoint, but
requests for them on my site returned 404. I tried to re-upload the
original to the same path but it appears since it had the same content
some optimization was preventing it from restoring the deleted file,
because it continued to 404.

Uploading the same file to a different path inside .well-known worked,
and then uploading a file with different content to the original path
worked.

This commit changes the ACME cleanup code to only remove the
.well-known/acme-challenge directory that it uses, and leaves any other
files alone.